+++ source/core/loaders/ObjectLoader.js
@@ -1,7 +1,6 @@
 import {AnimationClip, Bone, BufferAttribute, Color, DefaultLoadingManager, FileLoader, Fog, FogExp2, LOD, Line, LineLoop, LineSegments, Points, Skeleton} from "three";
 import {AmbientLight} from "../objects/lights/AmbientLight.js";
 import {AudioEmitter} from "../objects/audio/AudioEmitter.js";
-import {CubeCamera} from "../objects/cameras/CubeCamera.js";
 import {DirectionalLight} from "../objects/lights/DirectionalLight.js";
 import {EffectComposer} from "../postprocessing/EffectComposer.js";
 import {Group} from "../objects/misc/Group.js";
@@ -547,10 +546,6 @@ ObjectLoader.prototype.parseObject = function(data)
 				object.intensity = data.intensity;
 			}
 
-			break;
-
-		case "CubeCamera":
-			object = new CubeCamera(data.near, data.far, data.resolution, data.autoUpdate);
 			break;
 
 		case "OrbitControls":